Turn off scrollbars early in startup to avoid window width weirdness.
Sometimes the scrollbars would disappear, but the space they used to occupy
wasn't reclaimed by the window inside the frame -- both the outside frame and
the inner window remained the same width as when the scrollbars were enabled,
with unused space remaining where the scrollbars used to be.

This was on OS X 10.7.3 with GNU Emacs 24.1.50.2 (x86_64-apple-darwin11.3.0, NS
apple-appkit-1138.32) of 2012-05-04
(commit 74d659f8 of the git mirror git://git.savannah.gnu.org/emacs.git).