Enable whitespace-mode when editing, but not viewing, files.
And default to view mode when opening programming files.

This is a bit of an experiment. The idea is that I want whitespace-mode to
stop myself from committing whitespace in violation of the coding standards;
but those some coding standards are not rigorously adhered to throughout the
project and having whitespace-mode enabled all the time just fills my screen
with distracting red (even with the low-key pastelly red colour I've chosen).