Display "which function" in modeline before the list of active modes
This changes from:
  vc modes which-function
to:
  which-function vc modes

In an 80-character-wide frame, I usually couldn't see the current
function name (from which-function-mode) in the modeline. I rarely care
about the list of modes, so I pushed them to the end. vc's modeline
shows the current branch so it is definitely useful to keep visible, but
not as useful as the current function under point.