vc-annotate uses git blame's "-M" to detect moved/copied lines within a file
It already used "-C -C" to detect lines moved or copied from other files
that were modified in the same commit (the first "-C") and the commit
that creates the file (the second "-C"). If you give a third "-C" it
looks for copies from other files in any commit, but presumably that
will be really slow. Before I enable that I'll need to find some plugin
to display the annotate buffer incrementally (with each line showing the
most coarse "last changed" commit until more accurate information is
known).